+++ Character.cpp
@@ -50,7 +50,7 @@ Character::Character(const std::string& name, const std::string& race, int vital
 void Character::setName(const std::string& name)
 {
     std::string nameUpper = "";
-    for(int i = 0; i<name.size();i++)
+    for(size_t i = 0; i<name.size();i++)
     {
         if(std::isalpha(name[i]))
         {
